\name{DOALIAS}
\alias{DOALIAS}
\title{Map from DO alias to DO terms}
\description{
  DOALIAS is an R object that provides mapping
  from DO alias to DO terms
}
\details{
  Mappings were based on data provided by: Disease Ontology
  With a date stamp from the source of: 20150323 (sub_version 2806)


}
\references{
\url{http://do-wiki.nubic.northwestern.edu/index.php/Main_Page} 
%and \url{http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene} 
}
\keyword{datasets}

