\name{DOyulab.db}
\alias{DOyulab.db}
\alias{DO}
\title{Bioconductor annotation data package}
\description{

  Welcome to the DOyulab.db annotation Package.  The purpose of this package
  is to provide detailed information about the latest version of the
  Disease Ontology.  This package is updated biannually.
  
  You can learn what objects this package supports with the following command:
  
  \code{ls("package:DOyulab.db")} 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.
}
\keyword{datasets}

