\name{DOSYNONYM}
\alias{DOSYNONYM}
\title{Map from DO synonyms to DO terms}
\description{
    DOSYNONYM is an R object that provides mapping
    from DO synonyms to DO terms
}
\details{
    Mappings were based on data provided by: Disease Ontology
    With a date stamp from the source of: 20150323 (sub_version 2806)


}

\examples{
    # Convert the object to a list
    xx <- as.list(DOSYNONYM)
}

\references{
\url{http://do-wiki.nubic.northwestern.edu/index.php/Main_Page} 
%and \url{http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene} 
}


\keyword{datasets}

