\name{HDOALIAS}
\alias{HDOALIAS}
\title{Map from HDO alias to HDO terms}
\description{
    HDOALIAS is an R object that provides mapping
    from HDO alias to HDO terms
}
\details{
    Mappings were based on data provided by: Disease Ontology
    With a date stamp from the source of: 20220706


}

\examples{
    # Convert the object to a list
    xx <- as.list(HDOALIAS)
}


\references{
\url{http://hdo-wiki.nubic.northwestern.edu/index.php/Main_Page} 
%and \url{http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene} 
}
\keyword{datasets}

