\name{HDOGENE}
\alias{HDOGENE}
\title{Annotation of HDO Identifiers to gene ids}
\description{
    This data set gives mappings between HDO identifiers and gene ids.
}
\details{
    Mappings were based on data provided by: alliancegenome
    With a date stamp from the source of: 20240723
}

\examples{
    # Convert the object to a list
    xx <- as.list(HDOGENE)
}
\references{
\url{https://github.com/DiseaseOntology/HumanDiseaseOntology}
%and \url{https://fms.alliancegenome.org/download/DISEASE-ALLIANCE_HUMAN.tsv.gz}
}
\keyword{datasets}

