\name{HDOGENENCG}
\alias{HDOGENENCG}
\title{Annotation of NCG Identifiers to gene ids}
\description{
    This data set gives mappings between NCG identifiers and gene ids.
}
\details{
    Mappings were based on data provided by: NCG
    With a date stamp from the source of: 20240723
}

\examples{
    # Convert the object to a list
    xx <- as.list(HDOGENENCG)
}
\references{
\url{http://ncg.kcl.ac.uk/download.php}
}
\keyword{datasets}

