\name{HDOOBSOLETE}
\alias{HDOOBSOLETE}
\title{Annotation of HDO identifiers by terms defined by Disease Ontology
    Consortium and their status are obsolete}
\description{
This is an R object mapping HDO identifiers to the specific
terms in defined by Disease Ontology Consortium and their definition
are obsolete
}
\details{
    All the obsolete HDO terms that are collected in this index will
    no longer exist in other mapping objects.

    Mappings were based on data provided by: Disease Ontology
    With a date stamp from the source of: 220220706


}
\references{
\url{http://do-wiki.nubic.northwestern.edu/index.php/Main_Page}
%and \url{http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene}
}
\keyword{datasets}

