\name{HDOSYNONYM}
\alias{HDOSYNONYM}
\title{Map from HDO synonyms to HDO terms}
\description{
    HDOSYNONYM is an R object that provides mapping
    from HDO synonyms to HDO terms
}
\details{
    Mappings were based on data provided by: Disease Ontology
    With a date stamp from the source of: 20220706


}

\examples{
    # Convert the object to a list
    xx <- as.list(HDOSYNONYM)
}

\references{
\url{http://do-wiki.nubic.northwestern.edu/index.php/Main_Page} 
%and \url{http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene} 
}


\keyword{datasets}

